fix: update template names
