--- packages/openui-cli/src/templates/openui-chat/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "openui-chat",
-  "version": "0.1.1",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "eslint"
-  },
-  "dependencies": {
-    "@openuidev/react-ui": "latest",
-    "@openuidev/react-headless": "latest",
-    "@openuidev/react-lang": "latest",
-    "next": "16.1.6",
-    "openai": "^6.22.0",
-    "react": "19.2.3",
-    "react-dom": "19.2.3"
-  },
-  "devDependencies": {
-    "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "eslint": "^9",
-    "eslint-config-next": "16.1.6",
-    "tailwindcss": "^4",
-    "typescript": "^5"
-  }
-}
