--- packages/openui-cli/src/templates/openui-chat/src/app/api/chat/route.ts
@@ -1,31 +1,0 @@
-import { NextRequest } from "next/server";
-import OpenAI from "openai";
-
-const client = new OpenAI();
-
-export async function POST(req: NextRequest) {
-  try {
-    const { messages, systemPrompt } = await req.json();
-
-    const response = await client.chat.completions.create({
-      model: "gpt-5.2",
-      messages: [{ role: "system", content: systemPrompt }, ...messages],
-      stream: true,
-    });
-
-    return new Response(response.toReadableStream(), {
-      headers: {
-        "Content-Type": "text/event-stream",
-        "Cache-Control": "no-cache, no-transform",
-        Connection: "keep-alive",
-      },
-    });
-  } catch (err) {
-    console.error(err);
-    const message = err instanceof Error ? err.message : "Unknown error";
-    return new Response(JSON.stringify({ error: message }), {
-      status: 500,
-      headers: { "Content-Type": "application/json" },
-    });
-  }
-}
