--- packages/openui-cli/src/templates/openui-chat/src/app/globals.css
@@ -1,1 +1,0 @@
-@import "tailwindcss";
