+++ packages/openui-cli/src/templates/openui-cloud/.gitignore
@@ -36,9 +36,13 @@ yarn-error.log*
 # vercel
 .vercel
 
-# generated
-/src/generated/
+# local thread index (created at runtime)
+/.data/
 
 # typescript
 *.tsbuildinfo
 next-env.d.ts
+
+# vendored @openuidev/thesys tarball — obtain separately; never commit (the
+# package is distributed via the registry, not this repo).
+/vendor/
