+++ packages/openui-cli/src/templates/openui-cloud/README.md
@@ -1,0 +1,44 @@
+This is an [OpenUI](https://openui.com) Cloud project bootstrapped with [`openui-cli`](https://openui.com/docs/chat/quick-start).
+
+## Setup
+
+```bash
+cp .env.example .env.local   # fill THESYS_API_KEY and point the base URLs at your API
+```
+
+Required env: `THESYS_API_KEY`, `OPENUI_MODEL` (bare `provider/model`, e.g. `openai/gpt-5`), `DEMO_USER_ID`.
+
+## Getting Started
+
+First, run the development server:
+
+```bash
+npm run dev
+# or
+yarn dev
+# or
+pnpm dev
+# or
+bun dev
+```
+
+Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+
+You can start editing the page by modifying `src/app/api/route.ts` and improving your agent
+by adding system prompts or tools.
+
+## SDK packages
+
+- `@openuidev/thesys-server` — the server SDK (`artifactTool`,
+  `createResponsesInstructions`) used by the `/api/chat` route.
+- `@openuidev/thesys` — the React component library (`chatLibrary`, `Presentation`,
+  `Report`) used by the client page and artifact renderers.
+- `@openuidev/react-headless` / `@openuidev/react-ui` — the chat UI runtime
+  (`AgentInterface`, storage/stream contracts, `defineArtifactRenderer`).
+
+## Learn More
+
+To learn more about OpenUI, take a look at the following resources:
+
+- [OpenUI Documentation](https://openui.com/docs) - learn about OpenUI features and API.
+- [OpenUI GitHub repository](https://github.com/thesysdev/openui) - your feedback and contributions are welcome!
