+++ packages/openui-cli/src/templates/openui-cloud/eslint.config.mjs
