+++ packages/openui-cli/src/templates/openui-cloud/src/app/api/frontend-token/route.ts
