+++ packages/openui-cli/src/templates/openui-cloud/src/app/page.tsx
@@ -1,0 +1,79 @@
+"use client";
+import "@openuidev/react-ui/components.css";
+import "@openuidev/thesys/styles.css";
+
+import { useTheme } from "@/hooks/use-system-theme";
+import {
+  defineArtifactCategories,
+  openAIConversationMessageFormat,
+  openAIResponsesAdapter,
+  type ChatLLM,
+} from "@openuidev/react-headless";
+import { AgentInterface } from "@openuidev/react-ui";
+import {
+  chatLibrary,
+  presentationArtifactRenderer,
+  reportArtifactRenderer,
+  useOpenuiCloudStorage,
+} from "@openuidev/thesys";
+
+const { artifactRenderers, artifactCategories } = defineArtifactCategories([
+  { name: "Presentations", renderers: [presentationArtifactRenderer] },
+  { name: "Reports", renderers: [reportArtifactRenderer] },
+]);
+
+const llm: ChatLLM = {
+  send: async ({ threadId, messages, signal }) => {
+    const latest = messages.slice(-1);
+    return fetch("/api/chat", {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify({ threadId, input: openAIConversationMessageFormat.toApi(latest) }),
+      signal,
+    });
+  },
+  streamProtocol: openAIResponsesAdapter(),
+};
+
+export default function Page() {
+  const mode = useTheme();
+  const storage = useOpenuiCloudStorage({
+    token: "/api/frontend-token",
+    apiBaseUrl: "https://api.thesys.dev",
+    features: { artifact: true },
+  });
+
+  return (
+    <div className="h-screen w-screen overflow-hidden relative">
+      <AgentInterface
+        storage={storage}
+        llm={llm}
+        componentLibrary={chatLibrary}
+        artifactRenderers={artifactRenderers}
+        artifactCategories={artifactCategories}
+        agentName="OpenUI Cloud"
+        scrollVariant="always"
+        scrollOnLoad={false}
+        theme={{ mode }}
+        starters={[
+          {
+            displayText: "Summarize EV trends",
+            prompt: "In a few sentences, summarize the biggest EV market trends this quarter.",
+          },
+          {
+            displayText: "Pricing strategy tips",
+            prompt: "List five quick tips for pricing a new electric vehicle competitively.",
+          },
+          {
+            displayText: "Quarterly deck",
+            prompt: "Create a short presentation about our Q2 results with three slides.",
+          },
+          {
+            displayText: "Market report",
+            prompt: "Write a brief market-analysis report on the EV sector.",
+          },
+        ]}
+      />
+    </div>
+  );
+}
