+++ packages/openui-cli/src/templates/openui-cloud/src/hooks/use-system-theme.tsx
