+++ packages/openui-cli/src/templates/openui-cloud/src/lib/env.ts
