+++ packages/openui-cli/src/templates/openui-cloud/tsconfig.json
