+++ packages/openui-cli/src/templates/openui-self-hosted/.gitignore
@@ -36,13 +36,9 @@ yarn-error.log*
 # vercel
 .vercel
 
-# local thread index (created at runtime)
-/.data/
+# generated
+/src/generated/
 
 # typescript
 *.tsbuildinfo
 next-env.d.ts
-
-# vendored @openuidev/thesys tarball — obtain separately; never commit (the
-# package is distributed via the registry, not this repo).
-/vendor/
