+++ packages/openui-cli/src/templates/openui-self-hosted/README.md
@@ -1,12 +1,4 @@
-This is an [OpenUI](https://openui.com) Cloud project bootstrapped with [`openui-cli`](https://openui.com/docs/chat/quick-start).
-
-## Setup
-
-```bash
-cp .env.example .env.local   # fill THESYS_API_KEY and point the base URLs at your API
-```
-
-Required env: `THESYS_API_KEY`, `OPENUI_MODEL` (bare `provider/model`, e.g. `openai/gpt-5`), `DEMO_USER_ID`.
+This is an [OpenUI](https://openui.com) Agent Chat project bootstrapped with [`openui-cli`](https://openui.com/docs/chat/quick-start).
 
 ## Getting Started
 
@@ -26,15 +18,6 @@ Open [http://localhost:3000](http://localhost:3000) with your browser to see the
 
 You can start editing the page by modifying `src/app/api/route.ts` and improving your agent
 by adding system prompts or tools.
-
-## SDK packages
-
-- `@openuidev/thesys-server` — the server SDK (`artifactTool`,
-  `createResponsesInstructions`) used by the `/api/chat` route.
-- `@openuidev/thesys` — the React component library (`chatLibrary`, `Presentation`,
-  `Report`) used by the client page and artifact renderers.
-- `@openuidev/react-headless` / `@openuidev/react-ui` — the chat UI runtime
-  (`AgentInterface`, storage/stream contracts, `defineArtifactRenderer`).
 
 ## Learn More
