+++ packages/openui-cli/src/templates/openui-self-hosted/package.json
@@ -1,46 +1,21 @@
 {
-  "name": "openui-cloud",
-  "version": "0.1.0",
+  "name": "openui-self-hosted",
+  "version": "0.1.1",
   "private": true,
   "scripts": {
-    "dev": "next dev -p 3300",
+    "dev": "next dev",
     "build": "next build",
-    "start": "next start -p 3300",
-    "lint": "eslint",
-    "typecheck": "tsc --noEmit",
-    "test": "vitest run"
+    "start": "next start",
+    "lint": "eslint"
   },
   "dependencies": {
-    "@floating-ui/react-dom": "2.1.3",
-    "@openuidev/lang-core": "latest",
+    "@openuidev/react-ui": "latest",
     "@openuidev/react-headless": "latest",
     "@openuidev/react-lang": "latest",
-    "@openuidev/react-ui": "latest",
-    "@openuidev/thesys": "latest",
-    "@openuidev/thesys-server": "latest",
-    "@radix-ui/react-dialog": "1.1.15",
-    "@radix-ui/react-tooltip": "^1.2.0",
-    "@tanstack/react-table": "8.21.3",
-    "@tiptap/extension-placeholder": "2.27.2",
-    "@tiptap/react": "2.27.2",
-    "@tiptap/starter-kit": "2.27.2",
-    "clsx": "2.1.1",
-    "katex": "0.16.44",
-    "lodash": "4.17.21",
-    "lucide-react": "^0.575.0",
-    "mermaid": "11.15.0",
     "next": "16.1.6",
+    "openai": "^6.22.0",
     "react": "19.2.3",
-    "react-dom": "19.2.3",
-    "recharts": "2.15.4",
-    "rehype-katex": "7.0.1",
-    "remark-breaks": "4.0.0",
-    "remark-gfm": "4.0.1",
-    "remark-math": "6.0.0",
-    "tiny-invariant": "1.3.3",
-    "unist-util-visit": "5.1.0",
-    "zod": "^4.0.0",
-    "zustand": "^4.5.5"
+    "react-dom": "19.2.3"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4",
@@ -49,9 +24,7 @@
     "@types/react-dom": "^19",
     "eslint": "^9",
     "eslint-config-next": "16.1.6",
-    "openai": "^6.22.0",
     "tailwindcss": "^4",
-    "typescript": "^5",
-    "vitest": "^4.1.0"
+    "typescript": "^5"
   }
 }
