+++ packages/openui-cli/src/templates/openui-self-hosted/src/app/globals.css
@@ -1,12 +1,1 @@
 @import "tailwindcss";
-
-.openui-detailed-view-panel {
-  flex: 1 1 0% !important;
-  min-height: 0 !important;
-  display: flex !important;
-  flex-direction: column !important;
-}
-.openui-detailed-view-panel > .thesys-artifact-standalone {
-  flex: 1 1 0% !important;
-  min-height: 0 !important;
-}
