+++ packages/openui-cli/src/templates/openui-self-hosted/src/library.ts
